requirements.txt
torch
torchvision
opencv-python
streamlit
mlflow
fastapi
uvicorn
dvc
click==8.1.3
protobuf==3.20.*
fastapi>=0.68.0
uvicorn>=0.15.0
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0
python-multipart>=0.0.5
streamlit>=1.12.0
requests>=2.28.0
matplotlib>=3.6.0
pillow>=9.0.0
numpy>=1.23.0
